Update build info

+++ linux.sh
@@ -2,10 +2,16 @@
 # Script to download the latest x64 linux version of custom ffmpeg build for yt-dlp
 # The binary will be placed in the root dir of the app
 
-rm ffmpeg*
-wget "https://github.com/yt-dlp/FFmpeg-Builds/releases/latest/download/ffmpeg-n5.1-latest-linux64-gpl-5.1.tar.xz"
-tar xvf ffmpeg-n5.1-latest-linux64-gpl-5.1.tar.xz
-cp ffmpeg-n5.1-latest-linux64-gpl-5.1/bin/ffmpeg ffmpeg
+if ls ffmpeg* 1> /dev/null 2>&1; then
+    rm ffmpeg*
+fi
+# >> Check if curl is installed or nor
+if ! command -V curl > /dev/null 2>&1; then
+    echo "curl not installed, please install it and try again"
+    exit
+fi
+
+curl -SL --progress-bar "https://github.com/aandrew-me/ffmpeg-builds/releases/download/v6/ffmpeg_linux_amd64" -o ffmpeg
+
 chmod +x ffmpeg
-rm -rf ffmpeg-n5.1-latest-linux64-gpl-5.1
-rm ffmpeg-n5.1-latest-linux64-gpl-5.1.tar.xz
+
